Move router to gem
