+++ Gemfile
@@ -6,6 +6,8 @@ ruby '2.7.1'
 gem 'puma'
 gem 'rack'
 gem 'rubocop', require: false
+gem 'wabi_router', path: './router'
 
+gem 'pry'
 gem 'rack-test'
 gem 'rspec'
