+++ lib/wabi/base.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require './middleware/utils/cache'
-require './lib/wabi/router/base'
 require_relative 'parameters'
 
 module Wabi
@@ -23,13 +22,13 @@ module Wabi
     end
 
     def self.resources(resource_plural_name, except: [])
-      Router::Base
+      Wabi::Router
         .instance
         .add_resources_route(resource_plural_name, except)
     end
 
     def self.mount(path, app_class)
-      Router::Base
+      Wabi::Router
         .instance
         .add_mount_route(path, app_class)
     end
@@ -57,16 +56,17 @@ module Wabi
     private
 
     def resolve
-      @route = Router::Base.instance.find_route(@request.request_method, @request.path_info)
+      @route = Wabi::Router.instance.find_route(@request.request_method, @request.path_info)
       return NOT_FOUND_RESPONSE unless @route
 
       @response = Rack::Response[*generate_response]
       @response.finish
     end
 
+    # this method is related to router
     def generate_response
       case @route
-      when Wabi::Router::MountRoute, Wabi::Router::ResourcesRoute
+      when Wabi::MountRoute, Wabi::ResourcesRoute
         @route.response(@request.env)
       else
         route_response
@@ -84,7 +84,7 @@ module Wabi
     end
 
     def self.add_route(http_verb, path, block)
-      Router::Base
+      Wabi::Router
         .instance
         .add_route(http_verb, path, block)
     end
