+++ router/lib/wabi_router.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'wabi/router'
