+++ router/spec/spec_helper.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'bundler'
+Bundler.require
+
+require 'wabi/router'
