+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-require 'rack/test'
+require 'bundler'
+Bundler.require
 
 require './app'
 require './middleware/logger'
